chore(tests): move plugins from tests/parse_table_api.rs to plugin_collection
