+++ falco_plugin_tests/src/plugin_collection/extract/extra_fields.rs
@@ -1,0 +1,69 @@
+use crate::plugin_collection::tables::remaining_import_extra_fields::accessors::as_string::get_as_string;
+use crate::plugin_collection::tables::remaining_import_extra_fields::accessors::is_even::get_is_even;
+use crate::plugin_collection::tables::remaining_import_extra_fields::RemainingCounterImportTableWithExtraFields;
+use anyhow::Error;
+use falco_plugin::base::Plugin;
+use falco_plugin::event::events::types::EventType;
+use falco_plugin::event::events::types::EventType::PLUGINEVENT_E;
+use falco_plugin::extract::{field, ExtractFieldInfo, ExtractPlugin, ExtractRequest};
+use falco_plugin::static_plugin;
+use falco_plugin::tables::TablesInput;
+use std::ffi::{CStr, CString};
+
+struct ExtractExtraFields {
+    // reusing the table definition with the #[custom] annotations
+    // technically causes the fields to be added again, but we get
+    // the existing instances in that case
+    remaining_table: RemainingCounterImportTableWithExtraFields,
+}
+
+impl Plugin for ExtractExtraFields {
+    const NAME: &'static CStr = c"extract_extra_fields";
+    const PLUGIN_VERSION: &'static CStr = c"0.0.0";
+    const DESCRIPTION: &'static CStr = c"test plugin";
+    const CONTACT: &'static CStr = c"rust@localdomain.pl";
+    type ConfigType = ();
+
+    fn new(input: Option<&TablesInput>, _config: Self::ConfigType) -> Result<Self, Error> {
+        let input = input.ok_or_else(|| anyhow::anyhow!("did not get table input"))?;
+        let remaining_table = input.get_table(c"remaining")?;
+
+        Ok(Self { remaining_table })
+    }
+}
+
+impl ExtractExtraFields {
+    fn extract_is_even(&mut self, req: ExtractRequest<Self>) -> Result<u64, Error> {
+        let event_num = req.event.event_number() as u64;
+
+        let entry = self
+            .remaining_table
+            .get_entry(req.table_reader, &event_num)?;
+        let is_even = entry.get_is_even(req.table_reader)?;
+
+        Ok(is_even.into())
+    }
+
+    fn extract_string_rep(&mut self, req: ExtractRequest<Self>) -> Result<CString, Error> {
+        let event_num = req.event.event_number() as u64;
+
+        let entry = self
+            .remaining_table
+            .get_entry(req.table_reader, &event_num)?;
+        let string_rep = entry.get_as_string(req.table_reader)?;
+
+        Ok(CString::from(string_rep))
+    }
+}
+
+impl ExtractPlugin for ExtractExtraFields {
+    const EVENT_TYPES: &'static [EventType] = &[PLUGINEVENT_E];
+    const EVENT_SOURCES: &'static [&'static str] = &["countdown"];
+    type ExtractContext = ();
+    const EXTRACT_FIELDS: &'static [ExtractFieldInfo<Self>] = &[
+        field("countdown.is_even", &Self::extract_is_even),
+        field("countdown.as_string", &Self::extract_string_rep),
+    ];
+}
+
+static_plugin!(pub EXTRACT_EXTRA_FIELDS_API = ExtractExtraFields);
