+++ falco_plugin_tests/src/plugin_collection/extract/mod.rs
@@ -1,2 +1,3 @@
+pub mod extra_fields;
 pub mod remaining_from_table;
 pub mod remaining_from_table_runtime;
